fix: address PR review — decay_weight=0 fallback and CLI parseFloat
- Base sort now falls back to updated_at DESC when decay_weight=0,
  preserving backward-compatible ordering for the no-task path
- --decay-half-life CLI option changed from parseInt to parseFloat
  to prevent NaN from fractional inputs like '0.5'
- Added defensive guard in temporalDecay() for halfLifeDays <= 0
- Added tests for decay_weight=0 updated_at ordering and halfLife<=0

+++ src/cli/mk.ts
@@ -139,7 +139,7 @@ program
   .option('--paths <paths...>', 'Scope paths to match')
   .option('--types <types...>', 'Filter by atom type')
   .option('--max-tokens <n>', 'Token budget', parseInt)
-  .option('--decay-half-life <days>', 'Half-life for temporal decay in days (default: 30)', parseInt)
+  .option('--decay-half-life <days>', 'Half-life for temporal decay in days (default: 30)', parseFloat)
   .option('--decay-weight <n>', 'Weight of recency in scoring, 0-1 (default: 0.2)', parseFloat)
   .option('--include-episodes', 'Include EPISODES/ session summaries in context bundle')
   .action((opts: {
